api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
supabase==2.0.3
python-dotenv==1.0.0
langgraph==0.0.10
langchain==0.1.0
langchain-openai==0.0.2
redis==5.0.1
websockets==12.0
python-multipart==0.0.6
